requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# ScaleDown API - Core compression and optimizers
scaledown[haste,semantic]==0.1.4

# AI Models (Optional but recommended)
openai==1.54.3

# Data Processing
pandas==2.2.3
numpy==2.0.2

# HTTP Requests
requests==2.32.3
httpx==0.27.2

# Security & Validation
pydantic==2.9.2
python-multipart==0.0.12

# Utilities
python-dateutil==2.9.0

# Logging & Monitoring
loguru==0.7.2

# Testing (Development)
pytest==8.3.3
pytest-asyncio==0.24.0

# PDF Generation
reportlab==4.0.7

# Code Quality (Development)
black==24.10.0
flake8==7.1.1
